+++ cmd/bosun/expr/tsdbs/clients.go
@@ -1,4 +1,4 @@
-package expr
+package tsdbs
 
 import (
 	ainsightsmgmt "github.com/Azure/azure-sdk-for-go/services/appinsights/mgmt/2015-05-01/insights"
